backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0
psycopg2-binary
alembic
python-jose[cryptography]
bcrypt==4.0.1
passlib==1.7.4
slowapi
pydantic>=2.0
pydantic-settings
pydantic[email]
httpx
pytest
pytest-asyncio
pytest-cov
structlog
python-multipart
python-dotenv
